git-svn-id: http://svn.bramp.net:8080/svn/PhD/Work/userspace/threadnetperf@2023 e02701c3-087c-644a-a2a3-551e05f81084

+++ print.c
@@ -62,7 +62,7 @@ void print_results( const struct settings *settings, int core, struct stats *sta
 	pthread_mutex_unlock( &printf_mutex );
 }
 
-void print_stats(unsigned long long sum,unsigned long long sumsquare,unsigned long long mean,float variance) {
+void print_stats(unsigned long long sum, unsigned long long sumsquare, unsigned long long mean, double variance) {
 	printf("sum %llu sumsquare %llu mean %llu variance %f\n",sum, sumsquare, mean, variance);
 }
